+++ tealr/src/teal_multivalue.rs
@@ -1,6 +1,6 @@
 use std::borrow::Cow;
 
-use crate::teal_data::TypeRepresentation;
+use crate::{Direction, TypeName};
 
 ///Represents a type
 #[derive(Debug, PartialEq, Eq, Clone)]
@@ -12,16 +12,16 @@ impl TealType {
     fn new(name: Cow<'static, str>, is_external: bool) -> Self {
         Self { name, is_external }
     }
-    ///generates a [TealType](crate::TealType) based on a type implementing [TealData](crate::TealData).
+    ///generates a [TealType](crate::TealType) based on a type implementing [TealData](crate::rlu::TealData).
     ///```
-    ///# use tealr::TealType;
-    ///let nummeric_i8 = TealType::from::<i8>();
-    ///let nummeric_float = TealType::from::<f32>();
+    ///# use tealr::{Direction,TealType};
+    ///let numeric_i8 = TealType::from::<i8>(Direction::ToLua);
+    ///let numeric_float = TealType::from::<f32>(Direction::ToLua);
     /// //both i8 and f32 become a "number" in lua/teal. As such, their TealTypes are equal.
-    ///assert_eq!(nummeric_i8,nummeric_float)
+    ///assert_eq!(numeric_i8,numeric_float)
     ///```
-    pub fn from<A: TypeRepresentation>() -> Self {
-        Self::new(A::get_type_name(), A::is_external())
+    pub fn from<A: TypeName>(dir: Direction) -> Self {
+        Self::new(A::get_type_name(dir), A::is_external())
     }
 }
 
@@ -32,27 +32,27 @@ impl TealType {
 pub trait TealMultiValue {
     ///Gets the types contained in this collection.
     ///Order *IS* important.
-    fn get_types() -> Vec<TealType>;
+    fn get_types(dir: Direction) -> Vec<TealType>;
 }
 
 macro_rules! impl_teal_multi_value {
     () => (
         impl TealMultiValue for () {
-            fn get_types() -> Vec<TealType> {
-                vec![]
+            fn get_types(_:Direction) -> Vec<TealType> {
+                Vec::new()
             }
         }
     );
 
     ($($names:ident) +) => (
         impl<$($names,)* > TealMultiValue for ($($names,)*)
-            where $($names: TypeRepresentation,)*
+            where $($names: TypeName,)*
         {
             #[allow(unused_mut)]
             #[allow(non_snake_case)]
-            fn get_types() ->  Vec<TealType>{
+            fn get_types(dir:Direction) ->  Vec<TealType>{
                 let mut types = Vec::new();
-                $(types.push(TealType::from::<$names>());)*
+                $(types.push(TealType::from::<$names>(dir));)*
                 types
             }
         }
@@ -61,12 +61,12 @@ macro_rules! impl_teal_multi_value {
 
 impl<A> TealMultiValue for A
 where
-    A: TypeRepresentation,
+    A: TypeName,
 {
     #[allow(unused_mut)]
     #[allow(non_snake_case)]
-    fn get_types() -> Vec<TealType> {
-        vec![TealType::from::<A>()]
+    fn get_types(dir: Direction) -> Vec<TealType> {
+        vec![TealType::from::<A>(dir)]
     }
 }