+++ tealr_derive/CHANGES.md
@@ -6,10 +6,11 @@ All notable changes to this project are documented in this file.
 - [`0.2.0`](#020)
 - [`0.1.1`](#011)
 
-## upcomming
-
+## upcoming
+- Make the UserData derive macro a bit more hygienic
+- Automatically implement `TypeBody` when implementing `UserData`
 ## 0.2.0
-- Make derive macros bit more hygenic
+- Make derive macros bit more hygienic
 - Add derive macro for `TypeRepresentation`
 - Add derive macro that implements both `rlua::UserData` and `TypeRepresentation`